Updated the installation guide
The old guide was not working and tools used were old, so i updated the guide for BlissOS 14/15.

The guide is more simpler and is updated. The installer tools had issue so i replaced it with a more direct method and there is no need to open the EFI partition.  Following these method guarantees a working install without leaving windows or opening any partition.